Refactor FieldFinder to separate file (#846)
