+++ apps/mobile-app/android/app/src/main/java/net/aliasvault/app/autofill/models/FieldType.kt
@@ -1,0 +1,11 @@
+package net.aliasvault.app.autofill.models
+
+/**
+ * Field types that can be extracted from AssistStructure.
+ */
+enum class FieldType {
+    EMAIL,
+    USERNAME,
+    PASSWORD,
+    UNKNOWN
+}
